Fix: several warnings during compile.
This commit fixes most of them. There are a few that persist, and I
don't understand why, as the warnings reference code that exists as-is
in other variants, that don't cause any warnings during compile. Hmm.

Also some Makefile tweaks for my local build that will revert back once
I'm done messing with things.

+++ src/region.c
@@ -881,7 +881,7 @@ genericptr_t p2;
     int damage;
 
     reg = (NhRegion *) p1;
-    damage = (int) reg->arg;
+    damage = (size_t) reg->arg;
 
     /* If it was a thick cloud, it dissipates a little first */
     if (damage >= 5) {
@@ -903,7 +903,7 @@ genericptr_t p2;
     int dam;
 
     reg = (NhRegion *) p1;
-    dam = (int) reg->arg;
+    dam = (size_t) reg->arg;
     if (p2 == NULL) {		/* This means *YOU* Bozo! */
 	if (nonliving(youmonst.data) || Breathless)
 	    return FALSE;
